+++ src/viem/position/internal.ts
@@ -1,0 +1,2 @@
+export const waitForMs = (ms: number) =>
+  new Promise((resolve) => setTimeout(resolve, ms));
